Final version

+++ src/html/CssCopier.java
@@ -1,5 +1,7 @@
 package html;
 
+import main.Constants;
+
 import java.io.*;
 
 /**
@@ -8,22 +10,23 @@ import java.io.*;
  */
 public class CssCopier {
     private InputStream in = getClass().getResourceAsStream("/style.css");
-    private final String CSS_FILE_NAME = "style.css";
 
     /**
      * Copy the CSS file in the specified folder with a special name
-     * @param pathDestination the path to the folder
+     * @param pathDestination to pathDestination
      */
     public void copy(String pathDestination) {
         try {
-            File cssFile = new File(pathDestination + CSS_FILE_NAME);
-            cssFile.createNewFile();
-            FileWriter out = new FileWriter(cssFile);
-            while (in.available() > 0) {
-                out.write(in.read());
+            File cssFile = new File(pathDestination + Constants.CSS_FILE_NAME);
+            if (!cssFile.exists()) {
+                cssFile.createNewFile();
+                FileWriter out = new FileWriter(cssFile);
+                while (in.available() > 0) {
+                    out.write(in.read());
+                }
+                in.close();
+                out.close();
             }
-            in.close();
-            out.close();
         } catch (FileNotFoundException e) {
             e.printStackTrace();
         } catch (IOException e) {
